move ci-adapters methods into the base class instead of a template one
